% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\docType{class}
\name{pai}
\alias{pai}
\title{An R6 class for pesticidal active ingredients and associated data}
\format{
An \code{\link{R6Class}} generator object
}
\description{
The class is initialised with an identifier which is generally an ISO common name.
Additional chemical information is retrieved from the internet if available.
}
\examples{
# On Travis, we get a certificate validation error,
# likely because the system (xenial) is so old,
# therefore don't run this example on Travis
if (Sys.getenv("TRAVIS") == "") {

atr <- pai$new("atrazine")
print(atr)
if (!is.null(atr$Picture)) {
  plot(atr)
}

}
}
\section{Super class}{
\code{\link[chents:chent]{chents::chent}} -> \code{pai}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{iso}}{ISO common name according to ISO 1750 as retreived from pesticidecompendium.bcpc.org}

\item{\code{bcpc}}{List of information retrieved from pesticidecompendium.bcpc.org}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pai-new}{\code{pai$new()}}
\item \href{#method-pai-clone}{\code{pai$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_PUF"><a href='../../chents/html/chent.html#method-chent-add_PUF'><code>chents::chent$add_PUF()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_TP"><a href='../../chents/html/chent.html#method-chent-add_TP'><code>chents::chent$add_TP()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_cwsat"><a href='../../chents/html/chent.html#method-chent-add_cwsat'><code>chents::chent$add_cwsat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_p0"><a href='../../chents/html/chent.html#method-chent-add_p0'><code>chents::chent$add_p0()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_soil_degradation"><a href='../../chents/html/chent.html#method-chent-add_soil_degradation'><code>chents::chent$add_soil_degradation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_soil_ff"><a href='../../chents/html/chent.html#method-chent-add_soil_ff'><code>chents::chent$add_soil_ff()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_soil_sorption"><a href='../../chents/html/chent.html#method-chent-add_soil_sorption'><code>chents::chent$add_soil_sorption()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="add_transformation"><a href='../../chents/html/chent.html#method-chent-add_transformation'><code>chents::chent$add_transformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="emf"><a href='../../chents/html/chent.html#method-chent-emf'><code>chents::chent$emf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="get_chyaml"><a href='../../chents/html/chent.html#method-chent-get_chyaml'><code>chents::chent$get_chyaml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="get_pubchem"><a href='../../chents/html/chent.html#method-chent-get_pubchem'><code>chents::chent$get_pubchem()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="get_rdkit"><a href='../../chents/html/chent.html#method-chent-get_rdkit'><code>chents::chent$get_rdkit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="pdf"><a href='../../chents/html/chent.html#method-chent-pdf'><code>chents::chent$pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="png"><a href='../../chents/html/chent.html#method-chent-png'><code>chents::chent$png()</code></a></span></li>
<li><span class="pkg-link" data-pkg="chents" data-topic="chent" data-id="try_pubchem"><a href='../../chents/html/chent.html#method-chent-try_pubchem'><code>chents::chent$try_pubchem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pai-new"></a>}}
\if{latex}{\out{\hypertarget{method-pai-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pai$new(
  iso,
  identifier = iso,
  smiles = NULL,
  inchikey = NULL,
  bcpc = TRUE,
  pubchem = TRUE,
  pubchem_from = "auto",
  rdkit = TRUE,
  template = NULL,
  chyaml = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pai-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pai-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pai$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
