% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\name{plot.chent}
\alias{plot.chent}
\title{Plot method for chent objects}
\usage{
\method{plot}{chent}(x, ...)
}
\arguments{
\item{x}{The chent object to be plotted}

\item{...}{Further arguments passed to \link[grImport:grid.picture]{grImport::grid.picture}}
}
\description{
Plot method for chent objects
}
\examples{
caffeine <- chent$new("caffeine")
print(caffeine)
if (!is.null(caffeine$Picture)) {
  plot(caffeine)
}
}
