% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\name{print.chent}
\alias{print.chent}
\title{Printing method for chent objects}
\usage{
\method{print}{chent}(x, ...)
}
\arguments{
\item{x}{The chent object to be printed}

\item{...}{Further arguments for compatibility with the S3 method}
}
\description{
Printing method for chent objects
}

