% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\name{print.pai}
\alias{print.pai}
\title{Printing method for pai objects (pesticidal active ingredients)}
\usage{
\method{print}{pai}(x, ...)
}
\arguments{
\item{x}{The chent object to be printed}

\item{...}{Further arguments for compatibility with the S3 method}
}
\description{
Printing method for pai objects (pesticidal active ingredients)
}
