% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\name{print.ppp}
\alias{print.ppp}
\title{Printing method for ppp objects (plant protection products)}
\usage{
\method{print}{ppp}(x, ...)
}
\arguments{
\item{x}{The chent object to be printed}

\item{...}{Further arguments for compatibility with the S3 method}
}
\description{
Printing method for ppp objects (plant protection products)
}
