% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chent.R
\docType{class}
\name{chent}
\alias{chent}
\title{An R6 class for chemical entities with associated data}
\format{An \code{\link{R6Class}} generator object}
\usage{
chent
}
\description{
The class is initialised with an identifier. Chemical information is retrieved from
the internet. Additionally, it can be generated using RDKit if RDKit and its
python bindings are installed.
}
\section{Fields}{

\describe{
\item{\code{identifier}}{The identifier that was used to initiate the object, with attribute 'source'}

\item{\code{inchikey}}{InChI Key, with attribute 'source'}

\item{\code{smiles}}{SMILES code, with attribute 'source'}

\item{\code{mw}}{Molecular weight, with attribute 'source'}

\item{\code{pubchem}}{List of information retreived from PubChem}

\item{\code{rdkit}}{List of information obtained with RDKit}

\item{\code{Picture}}{Graph as a \code{\link{picture}} object obtained using grImport}
}}
\examples{
oct <- chent$new("1-octanol", smiles = "CCCCCCCCO")
oct$try_pubchem()
print(oct)
plot(oct)
caffeine <- chent$new("caffeine", source = "pubchem")
print(caffeine)
caffeine$get_rdkit()
plot(caffeine)
}
\keyword{data}

