% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chent.R
\docType{class}
\name{chent}
\alias{chent}
\title{An R6 class for chemical entities with associated data}
\format{An \code{\link{R6Class}} generator object}
\usage{
chent
}
\description{
The class is initialised with an identifier. Chemical information is retrieved
from the internet.
}
\section{Fields}{

\describe{
\item{\code{identifier}}{The identifier that was used to initiate the object, with attribute 'source'}

\item{\code{inchikey}}{InChI Key, with attribute 'source'}

\item{\code{smiles}}{SMILES code, with attribute 'source'}

\item{\code{mw}}{Molecular weight, with attribute 'source'}

\item{\code{pubchem}}{List of information retreived from PubChem}

\item{\code{rdkit}}{List of information obtained with RDKit}
}}
\examples{
oct <- chent$new("1-octanol", smiles = "CCCCCCCCO")
oct$try_pubchem()
print(oct)
}
\keyword{data}

