% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\docType{class}
\name{chent}
\alias{chent}
\title{An R6 class for chemical entities with associated data}
\format{An \code{\link{R6Class}} generator object}
\usage{
chent
}
\description{
The class is initialised with an identifier. Chemical information is retrieved from
the internet. Additionally, it can be generated using RDKit if RDKit and its
python bindings are installed and configured for use with PythonInR.
}
\section{Fields}{

\describe{
\item{\code{identifier}}{The identifier that was used to initiate the object, with attribute 'source'}

\item{\code{inchikey}}{InChI Key, with attribute 'source'}

\item{\code{smiles}}{SMILES code, with attribute 'source'}

\item{\code{mw}}{Molecular weight, with attribute 'source'}

\item{\code{pubchem}}{List of information retreived from PubChem}

\item{\code{rdkit}}{List of information obtained with RDKit, if installed and
configured for use with PythonInR}

\item{\code{Picture}}{Graph as a \code{\link{picture}} object obtained using grImport}

\item{\code{chyaml}}{List of information obtained from a YAML file}

\item{\code{degradation}}{List of degradation endpoints}
}}
\examples{
oct <- chent$new("1-octanol", smiles = "CCCCCCCCO")
print(oct)
plot(oct)
caffeine <- chent$new("caffeine")
print(caffeine)
plot(caffeine)
}
\keyword{data}

