% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chent.R
\docType{class}
\name{pai}
\alias{pai}
\title{An R6 class for pesticidal active ingredients and associated data}
\format{An \code{\link{R6Class}} generator object}
\usage{
pai
}
\description{
The class is initialised with an identifier which is generally an ISO common name.
Additional chemical information is retrieved from the internet.
}
\section{Fields}{

\describe{
\item{\code{iso}}{ISO common name according to ISO 1750 as retreived from www.alanwood.net/pesticides}

\item{\code{alanwood}}{List of information retreived from www.alanwood.net/pesticides}
}}
\examples{
atr <- pai$new("atrazine")
print(atr)
}
\keyword{data}

