% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chent.R
\docType{class}
\name{pp}
\alias{pp}
\title{R6 class for holding a product with at least one active ingredient}
\format{An \code{\link{R6Class}} generator object.}
\usage{
pp
}
\description{
An R6 class for holding information about a product with at least one active ingredient
}
\section{Fields}{

\describe{
\item{\code{name}}{The name of the product}

\item{\code{ais}}{A list of active ingredients}

\item{\code{concentrations}}{The concentration of the ais}

\item{\code{concentration_units}}{Defaults to g/L}
}}
\keyword{data}

